% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_image.R
\name{rotate_image}
\alias{rotate_image}
\title{Rotate an image counter-clockwise}
\usage{
rotate_image(img, rot_num = 1)
}
\arguments{
\item{img}{image loaded via imager library}

\item{rot_num}{integer}
}
\value{
array of image
}
\description{
rotate_image rotates an image counter-clockwise for a given number of times.
}
\examples{
rotate_image(imager::boats, 3)
}
